refactor: 브로드캐스트 응답과 분리하기 위해 중첩 record로 변경
