+++ src/main/java/umc/cockple/demo/domain/chat/service/ChatWebSocketService.java
@@ -8,8 +8,8 @@ import umc.cockple.demo.domain.chat.converter.ChatConverter;
 import umc.cockple.demo.domain.chat.domain.*;
 import umc.cockple.demo.domain.chat.dto.MemberConnectionInfo;
 import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO;
-import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.FileInfo;
-import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.ImageInfo;
+import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.Request.FileInfo;
+import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.Request.ImageInfo;
 import umc.cockple.demo.domain.chat.enums.ChatRoomType;
 import umc.cockple.demo.domain.chat.enums.MessageType;
 import umc.cockple.demo.domain.chat.exception.ChatErrorCode;
